+++ lib/weber/helper/partial.ex
@@ -6,11 +6,17 @@ defmodule Weber.Helper.Partial do
     ## Example
 
     # Generates <div>this is a partial file</div>
-    <%= render "partial" %>
+    <%= render "Partial" %>
 
   """
 
-  def render(file) do
-    Path.expand("_" <> file <> ".html", Weber.Path.__root__ <> "/lib/views/partials") |> EEx.eval_file()
+  def render(file, bindings \\ []) do
+  	case bindings do
+  	  [] ->
+        Path.expand(file <> ".html", Weber.Path.__root__ <> "/lib/views/partials") |> EEx.eval_file()
+  	  _ ->
+  	  	partial = Module.concat(Elixir.Views.Partials, file)
+  	  	partial.render_template(bindings)
+  	end
   end
 end