+++ lib/weber/helper/resource_helper.ex
@@ -131,5 +131,5 @@ defmodule Weber.Helper.ResourceHelper do
   def audio(src, html_options) do
     tag(:audio, :lists.append([src: src], html_options))
   end
-
+  
 end