+++ lib/weber/utils/utils.ex
@@ -48,7 +48,7 @@ defmodule Weber.Utils do
   Collect all Helpers imports.
   """
   def add_helpers_imports(view_content) do
-    "<% import Weber.Helper.Html %>" <> "<% import Weber.Helper.Partial %>" <> "<% import Weber.Helper.IncludeView %>" <> 
+    "<% import Weber.Helper.Html %>" <> "<% import Weber.Helper.Partial %>" <> 
     "<% import Weber.Helper.ResourceHelper %>" <> "<% import Weber.I18n %>" <> view_content
   end